Refactor Integration Tests to make them less verbose
